pyproject.toml
[project]
name = "risk-analyzer"
version = "0.1.0"
description = "Straightforward LangGraph-based risk analyzer for Joget folios"
authors = [{ name = "JCTekDev", email = "jc.tek@yahoo.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.26",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
    "langchain-core>=0.2",
    "langgraph>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.setuptools.packages.find]
where = ["src"]
